Refactor the base http types
